+++ CHANGELOG.md
@@ -5,7 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-### [v0.3.0]
+## [Unreleased]
+
+### Added
+
+- Functionality from by package `bydi` to setup paths for tests and
+  other testing libraries was moved here under `dinghy-rope`.
+
+### Changed
+
+- Directory `scripts` was renamed to `src`. It is set by
+  `DINGHY_SRC_DIR`.
+
+## [v0.3.0]
 
 ### Added